+++ test.svelte
@@ -24,15 +24,23 @@
   type Extract = (item: Item) => string;
 
   const extract: Extract = (item) => item.state;
+
+  type DisableOrFilter = (item: Item) => boolean;
+
+  const disable: DisableOrFilter = (item) => item.state.length > 10;
+  const filter: DisableOrFilter = (item) => item.state.length < 4;
 </script>
 
 <Typeahead
   {extract}
+  {disable}
+  {filter}
   autocapitalize={false + ''}
   placeholder="#{4}"
   autofocus
   hideLabel
   focusAfterSelect
+  inputAfterSelect="keep"
   debounce={800}
   {data}
   on:select={(e) => {
@@ -44,7 +52,8 @@
   }}
   bind:results
   let:result
-  let:index>
+  let:index
+>
   {@html result.string}
   {index}
   {result.score}