+++ CHANGELOG.md
@@ -5,7 +5,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.0.0](https://github.com/metonym/svelte-typeahead/releases/tag/v1.0.0) - 2020-12-31
+## [2.3.0](https://github.com/metonym/svelte-typeahead/releases/tag/v2.3.0) - 2021-02-21
+
+**Features**
+
+- add `disable`, `filter` props to disable and filter items from the result set
+
+**Fixes**
+
+- bind the input element reference correctly to fix focusing behavior
+- don't pass the Typeahead id to Search
+
+## [2.2.0](https://github.com/metonym/svelte-typeahead/releases/tag/v2.2.0) - 2021-02-20
+
+**Features**
+
+- add `inputAfterSelect` prop to allow user to preserve or clear the input field after selecting a result; possible values are `"update" | "clear" | "keep"` (default is `"update"`)
+- add searched value to dispatched "select" event detail (`e.detail.searched`)
+
+## [2.1.0](https://github.com/metonym/svelte-typeahead/releases/tag/v2.1.0) - 2021-02-20
+
+**Features**
+
+- include `original` item and `originalIndex` in dispatched "select" event
+
+## [2.0.0](https://github.com/metonym/svelte-typeahead/releases/tag/v2.0.0) - 2020-12-31
 
 **Breaking Changes**